webapp/requirements.txt
fastapi
uvicorn
Pillow
torch
numpy
scikit-image
opencv-python
matplotlib